diff: narrow ints join the op matrix
u8/i8/u16/i16 get the full differential treatment: the ten int BinOps
over width-boundary, wrap-witness and sign-probe inputs (MIN/-1
included — defined at narrow width, unlike i32), Neg/BitNot, the six
compares with signedness pinned by 0x80-class operands, the complete
narrow<->wide/float cast lane, wide-const-into-narrow-op rows, and
256-element ramps per type so index > 0 exercises the native 1-/2-byte
stride. The dispatcher grows the narrow typed-field arms and dispatches
one quark per element; the WGSL browser audit skips narrow rows with
reason (WGSL 1.0 has no 8-/16-bit storage).

First exercise findings: the CPU reference zero-extended i8 loads —
every sign-sensitive i8 op read -1 as 255; fixed by sign-extending
through i8 at both read sites. Metal then passes all 982 narrow rows
bit-exact. Excluded with measured witnesses until their contracts are
settled: narrow Rotl/Rotr (the reference rotates at 32-bit, Metal and
SPIR-V at native width, and narrow Rotr does not even compile on MSL —
ambiguous rotate overloads), narrow SatAdd (32-bit saturation vs the
narrow clamp), and negative signed-narrow -> u64 casts (source-width
masking vs C sign-extension, the same gap i32 -> u64 has always had,
unexercised).

+++ crates/gpu/quanta-core/src/driver/cpu/value.rs
@@ -149,7 +149,8 @@ pub(super) fn read_scalar_at_offset(buf: &[u8], offset: usize, ty: &ScalarType) 
         ScalarType::U16 => Value::U32(u16::from_le_bytes(bytes.try_into().unwrap()) as u32),
         ScalarType::I16 => Value::I32(i16::from_le_bytes(bytes.try_into().unwrap()) as i32),
         ScalarType::U8 => Value::U32(bytes[0] as u32),
-        ScalarType::I8 => Value::I32(bytes[0] as i32),
+        // Sign-extend through i8 — a bare `bytes[0] as i32` zero-extends.
+        ScalarType::I8 => Value::I32(bytes[0] as i8 as i32),
         ScalarType::F16 => {
             let bits = u16::from_le_bytes(bytes.try_into().unwrap());
             Value::F32(f16_to_f32(bits))
@@ -195,7 +196,8 @@ pub(super) fn read_scalar(buf: &[u8], index: u32, ty: &ScalarType) -> Value {
         ScalarType::U16 => Value::U32(u16::from_le_bytes(bytes.try_into().unwrap()) as u32),
         ScalarType::I16 => Value::I32(i16::from_le_bytes(bytes.try_into().unwrap()) as i32),
         ScalarType::U8 => Value::U32(bytes[0] as u32),
-        ScalarType::I8 => Value::I32(bytes[0] as i32),
+        // Sign-extend through i8 — a bare `bytes[0] as i32` zero-extends.
+        ScalarType::I8 => Value::I32(bytes[0] as i8 as i32),
         ScalarType::F16 => {
             // f16: decode from u16 bits, expand to f32
             let bits = u16::from_le_bytes(bytes.try_into().unwrap());
